+++ mmros/include/mmros/tensorrt/plugin/grid_sampler/trt_grid_sampler_kernel.hpp
@@ -5,6 +5,8 @@
 
 #include <cuda_runtime.h>
 
+#include <cstdint>
+
 namespace mmros::plugin
 {
 /**
@@ -48,8 +50,8 @@ enum class GridSamplerPadding {
  */
 template <typename T>
 void grid_sample(
-  T * output, const T * input, const T * grid, int * output_dims, int * input_dims, int * grid_dims,
-  int nb_dims, GridSamplerInterpolation interp, GridSamplerPadding padding, bool align_corners,
-  cudaStream_t stream);
+  T * output, const T * input, const T * grid, int64_t * output_dims, int64_t * input_dims,
+  int64_t * grid_dims, int nb_dims, GridSamplerInterpolation interp, GridSamplerPadding padding,
+  bool align_corners, cudaStream_t stream);
 }  // namespace mmros::plugin
 #endif  // MMROS__TENSORRT__PLUGIN__GRID_SAMPLER__TRT_GRID_SAMPLER_KERNEL_HPP_
